move `print_result_c` into the trait
